+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.6.7](https://github.com/sunerpy/zuno/compare/v0.6.6...v0.6.7) (2026-09-02)
+
+
+### Bug Fixes
+
+* **engine:** reconcile only durably recorded work ([#90](https://github.com/sunerpy/zuno/issues/90)) ([6deec18](https://github.com/sunerpy/zuno/commit/6deec182d53cbb26745b695fa9eb379687c9bacc))
+* **tui:** discard unread input when leaving the terminal ([#91](https://github.com/sunerpy/zuno/issues/91)) ([b07cc1d](https://github.com/sunerpy/zuno/commit/b07cc1d4477555c0c6d3fda57269b5ad0052c690))
+
+## [0.6.6](https://github.com/sunerpy/zuno/compare/v0.6.5...v0.6.6) (2026-09-02)
+
+
+### Bug Fixes
+
+* **acp:** speed up session configuration changes ([#88](https://github.com/sunerpy/zuno/issues/88)) ([6f3340c](https://github.com/sunerpy/zuno/commit/6f3340c7c6eba92c8ff5ceea91eab9354e140006))
+
 ## [0.6.5](https://github.com/sunerpy/zuno/compare/v0.6.4...v0.6.5) (2026-09-02)